cpp: qtwidgets: Move gl rendering classes to versioned dir
