+++ cpp/lib/ome/qtwidgets/glsl/v110/GLLineShader2D.h
@@ -36,8 +36,8 @@
  * #L%
  */
 
-#ifndef OME_QTWIDGETS_GLSL_V110GLLINESHADER2D_H
-#define OME_QTWIDGETS_GLSL_V110GLLINESHADER2D_H
+#ifndef OME_QTWIDGETS_GLSL_V110_GLLINESHADER2D_H
+#define OME_QTWIDGETS_GLSL_V110_GLLINESHADER2D_H
 
 #include <QOpenGLShader>
 #include <QOpenGLBuffer>
@@ -162,7 +162,7 @@ namespace ome
   }
 }
 
-#endif // OME_QTWIDGETS_GLSL_V110GLLINESHADER2D_H
+#endif // OME_QTWIDGETS_GLSL_V110_GLLINESHADER2D_H
 
 /*
  * Local Variables:
